move and refactor manifest parser
