+++ backend/src/main/java/spectacular/backend/catalogues/CatalogueService.java
@@ -9,11 +9,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import spectacular.backend.api.model.Catalogue;
-import spectacular.backend.cataloguemanifest.CatalogueManifestParser;
 import spectacular.backend.cataloguemanifest.CatalogueManifestProvider;
 import spectacular.backend.cataloguemanifest.GetCatalogueManifestFileContentResult;
 import spectacular.backend.cataloguemanifest.catalogueentry.CatalogueEntryConfigurationResolver;
 import spectacular.backend.cataloguemanifest.interfaceentry.CatalogueInterfaceEntryConfigurationResolver;
+import spectacular.backend.cataloguemanifest.parse.CatalogueManifestParser;
 import spectacular.backend.common.CatalogueId;
 import spectacular.backend.interfaces.GetInterfaceFileContentsResult;
 import spectacular.backend.interfaces.InterfaceService;
