+++ backend/src/test/groovy/spectacular/backend/catalogues/CatalogueServiceTest.groovy
@@ -6,8 +6,8 @@ import spectacular.backend.api.model.GetInterfaceResult
 import spectacular.backend.api.model.SpecEvolutionSummary
 import spectacular.backend.cataloguemanifest.catalogueentry.CatalogueEntryConfigurationResolver
 import spectacular.backend.cataloguemanifest.interfaceentry.CatalogueInterfaceEntryConfigurationResolver
-import spectacular.backend.cataloguemanifest.CatalogueManifestContentItemParseResult
-import spectacular.backend.cataloguemanifest.CatalogueManifestParser
+import spectacular.backend.cataloguemanifest.parse.CatalogueManifestContentItemParseResult
+import spectacular.backend.cataloguemanifest.parse.CatalogueManifestParser
 import spectacular.backend.cataloguemanifest.CatalogueManifestProvider
 import spectacular.backend.cataloguemanifest.catalogueentry.GetCatalogueEntryConfigurationResult
 import spectacular.backend.cataloguemanifest.configurationitem.ConfigurationItemError
